Fixed genres

+++ src/main/java/tim2/cataloging/tim2/controller/GenreController.java
@@ -7,9 +7,11 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import tim2.cataloging.tim2.dto.GenreDto;
+import tim2.cataloging.tim2.model.Book;
 import tim2.cataloging.tim2.model.Genre;
 import tim2.cataloging.tim2.model.ROLE;
 import tim2.cataloging.tim2.model.User;
+import tim2.cataloging.tim2.service.BookService;
 import tim2.cataloging.tim2.service.GenreService;
 
 import java.util.ArrayList;
@@ -18,8 +20,12 @@ import java.util.List;
 @RestController
 @RequestMapping("/genres")
 public class GenreController {
+
     @Autowired
     private GenreService genreService;
+
+    @Autowired
+    private BookService bookService;
 
     //READ ALL
     @GetMapping("")
@@ -31,6 +37,7 @@ public class GenreController {
           GenreDto dto = new GenreDto(g);
           dtos.add(dto);
        }
+
        return ResponseEntity.ok(dtos);
     }
 
@@ -51,37 +58,58 @@ public class GenreController {
             return ResponseEntity.badRequest().body("Only admin can add genres");
 
         genreService.save(genre);
+
         return ResponseEntity.ok("Genre added successfully");
     }
 
     //UPDATE
     @PutMapping("/{id}")
-    public ResponseEntity<String> updateGenre(@PathVariable("id") Long id, @RequestBody Genre updatedGenre) {
-        try {
-            Genre existingGenre = genreService.findOne(id);
-            if (existingGenre == null) {
-                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Genre not found");
-            }
-            existingGenre.setName(updatedGenre.getName());
-            genreService.save(existingGenre); // Save the updated genre
+    public ResponseEntity<String> updateGenre(@PathVariable("id") Long id, @RequestBody Genre updatedGenre, HttpSession session) {
+        User loggedUser = (User) session.getAttribute("user");
+        if (loggedUser == null)
+            return ResponseEntity.badRequest().body("You must be logged in to update genres");
 
-            return ResponseEntity.ok("Genre " + id + " updated successfully");
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to update genre");
+        if(loggedUser.getRole() != ROLE.ADMIN)
+            return ResponseEntity.badRequest().body("Only admin can update genres");
+
+        Genre existingGenre = genreService.findOne(id);
+        if (existingGenre == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Genre not found");
         }
+
+        if (updatedGenre.getName() != null)
+            existingGenre.setName(updatedGenre.getName());
+
+        genreService.save(existingGenre); // Save the updated genre
+
+        return ResponseEntity.ok("Genre " + id + " updated successfully");
     }
+
     //DELETE
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteGenre(@PathVariable("id") Long id) {
-        try {
-            Genre genre = genreService.findOne(id);
-            if (genre == null) {
-                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Genre not found");
-            }
-            genreService.deleteById(id);
-            return ResponseEntity.ok("Genre " + id + " deleted successfully");
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to delete genre");
+    public ResponseEntity<String> deleteGenre(@PathVariable("id") Long id, HttpSession session) {
+        User loggedUser = (User) session.getAttribute("user");
+        if (loggedUser == null)
+            return ResponseEntity.badRequest().body("You must be logged in to delete genres");
+
+        if(loggedUser.getRole() != ROLE.ADMIN)
+            return ResponseEntity.badRequest().body("Only admin can delete genres");
+
+        Genre genre = genreService.findOne(id);
+        if (genre == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Genre not found");
         }
+
+        List<Book> books = bookService.findByGenreId(id);
+        if (books.size() > 0) {
+            for (Book b : books) {
+                b.setGenre(null);
+                bookService.save(b);
+            }
+        }
+
+        genreService.deleteById(id);
+
+        return ResponseEntity.ok("Genre " + id + " deleted successfully");
     }
 }
